Moving siepic_tools.extend.point_multiplication extension
