+++ klayout_dot_config/python/siepic_tools/_globals.py
@@ -1,0 +1,9 @@
+def check_numpy():
+    ''' Check if numpy is present
+    '''
+    try:
+        import numpy  # noqa
+    except ImportError:
+        return False
+    else:
+        return True
