+++ klayout_dot_config/python/siepic_tools/extend/__init__.py
@@ -1,0 +1,1 @@
+import siepic_tools.extend.point_multiplication  # noqa
