+++ klayout_dot_config/python/siepic_tools/utils/__init__.py
@@ -4,3 +4,6 @@ Utilities module for siepic_tools
 
 from .tech import get_technology_by_name  # noqa
 from .tech import register_siepic_technology  # noqa
+
+# The code here in utils require certain klayout extensions.
+import siepic_tools.extend  # noqa
